Refactor(Api): test api와 기존 api를 분리
